requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#

asn1crypto==1.5.1
    # via
    #   oscrypto
    #   snowflake-connector-python
astroid==3.0.1
    # via pylint
attrs==23.1.0
    # via
    #   jsonschema
    #   referencing
black==23.10.0
    # via -r requirements.in
certifi==2023.7.22
    # via
    #   requests
    #   snowflake-connector-python
cffi==1.16.0
    # via
    #   cryptography
    #   snowflake-connector-python
charset-normalizer==3.3.0
    # via
    #   requests
    #   snowflake-connector-python
click==8.1.7
    # via black
colorama==0.4.6
    # via -r requirements.in
cryptography==41.0.4
    # via
    #   pyopenssl
    #   snowflake-connector-python
dill==0.3.7
    # via pylint
exceptiongroup==1.1.3
    # via pytest
filelock==3.12.4
    # via snowflake-connector-python
greenlet==3.0.0
    # via sqlalchemy
idna==3.4
    # via
    #   requests
    #   snowflake-connector-python
iniconfig==2.0.0
    # via pytest
invoke==2.2.0
    # via -r requirements.in
isort==5.12.0
    # via pylint
jinja2==3.1.2
    # via -r requirements.in
jsonschema==4.19.1
    # via -r requirements.in
jsonschema-specifications==2023.7.1
    # via jsonschema
markupsafe==2.1.3
    # via jinja2
mccabe==0.7.0
    # via pylint
mypy-extensions==1.0.0
    # via black
nest-asyncio==1.5.8
    # via -r requirements.in
networkx==3.2
    # via -r requirements.in
numpy==1.26.1
    # via pandas
oscrypto==1.3.0
    # via snowflake-connector-python
packaging==23.2
    # via
    #   black
    #   pytest
    #   snowflake-connector-python
pandas==2.1.1
    # via -r requirements.in
pathspec==0.11.2
    # via black
platformdirs==3.11.0
    # via
    #   black
    #   pylint
    #   snowflake-connector-python
pluggy==1.3.0
    # via pytest
pycparser==2.21
    # via cffi
pycryptodomex==3.19.0
    # via snowflake-connector-python
pyjwt==2.8.0
    # via snowflake-connector-python
pylint==3.0.1
    # via -r requirements.in
pyopenssl==23.2.0
    # via snowflake-connector-python
pytest==7.4.2
    # via -r requirements.in
python-dateutil==2.8.2
    # via pandas
pytz==2023.3.post1
    # via
    #   pandas
    #   snowflake-connector-python
pyyaml==6.0.1
    # via -r requirements.in
referencing==0.30.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.31.0
    # via snowflake-connector-python
rpds-py==0.10.6
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via python-dateutil
snowflake-connector-python==3.3.1
    # via snowflake-sqlalchemy
snowflake-sqlalchemy==1.5.0
    # via -r requirements.in
sortedcontainers==2.4.0
    # via snowflake-connector-python
sqlalchemy==1.4.49
    # via
    #   -r requirements.in
    #   snowflake-sqlalchemy
tomli==2.0.1
    # via
    #   black
    #   pylint
    #   pytest
tomlkit==0.12.1
    # via
    #   pylint
    #   snowflake-connector-python
typing-extensions==4.8.0
    # via
    #   astroid
    #   black
    #   snowflake-connector-python
    #   sqlalchemy
tzdata==2023.3
    # via pandas
urllib3==1.26.18
    # via
    #   requests
    #   snowflake-connector-python
